requirements.txt
# Django-specific packages
Django==1.5.1
South==0.7.6
django-browserid==0.7.1
django_compressor==1.3
django-haystack==2.0.0
pyelasticsearch==0.5
django-model-utils==1.3.0
django-reversion==1.7
djangorestframework==2.2.5

BeautifulSoup<4.0
lxml==3.0.2
cssselect==0.8
PIL==1.1.7
psycopg2==2.4.6

isodate==0.4.9
pytz==2012h
refine-client==0.2.1
smartypants==1.6.0.3

ipython==0.13.1
